# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": " App",
        "version": "v1.0.0"
    },
    "paths": {
...
        "/api/shop": {
            "get": {
                "tags": [
                    "Shop \u5546\u5e97\u7c7bApi"
                ],
                "summary": "\u6211\u7684\u5546\u5e97",
                "description": "\u5546\u5e97\u5217\u8868",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content_type",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            },
            "post": {
                "tags": [
                    "Shop \u5546\u5e97\u7c7bApi"
                ],
                "summary": "\u521b\u5efa\u6211\u7684\u5546\u5e97",
                "description": "\u521b\u5efa\u5546\u5e97",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content_type",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "\u5546\u5e97\u540d\u79f0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "logo",
                        "in": "formData",
                        "description": "\u7528\u6237\u5934\u50cf",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "location",
                        "in": "formData",
                        "description": "\u5e97\u94fa\u4f4d\u7f6e",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "\u5e97\u94fa\u7c7b\u578b",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "\u5546\u5e97\u63cf\u8ff0",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "open_at",
                        "in": "formData",
                        "description": "\u5f00\u95e8\u8425\u4e1a\u65f6\u95f4",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "close_at",
                        "in": "formData",
                        "description": "\u5173\u95e8\u6253\u70ca\u65f6\u95f4",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            }
        },
        "/api/shop/{id}": {
            "get": {
                "tags": [
                    "Shop \u5546\u5e97\u7c7bApi"
                ],
                "summary": "\u5546\u5e97\u8be6\u60c5",
                "description": "\u5546\u5e97\u8be6\u60c5",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content_type",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            },
            "put": {
                "tags": [
                    "Shop \u5546\u5e97\u7c7bApi"
                ],
                "summary": "\u66f4\u65b0\u5546\u5e97\u4fe1\u606f",
                "description": "\u66f4\u65b0\u5546\u5e97\u4fe1\u606f",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content_type",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "\u5546\u5e97\u540d\u79f0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "logo",
                        "in": "formData",
                        "description": "\u7528\u6237\u5934\u50cf",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "location",
                        "in": "formData",
                        "description": "\u5e97\u94fa\u4f4d\u7f6e",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "\u5e97\u94fa\u7c7b\u578b",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "\u5546\u5e97\u63cf\u8ff0",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "open_at",
                        "in": "formData",
                        "description": "\u5f00\u95e8\u8425\u4e1a\u65f6\u95f4",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "close_at",
                        "in": "formData",
                        "description": "\u5173\u95e8\u6253\u70ca\u65f6\u95f4",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Shop \u5546\u5e97\u7c7bApi"
                ],
                "summary": "\u5220\u9664\u5546\u5e97",
                "description": "\u5220\u9664\u5546\u5e97",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content_type",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            }
        },
        "/api/shop/certify": {
            "get": {
                "tags": [
                    "Shop \u5546\u5e97\u7c7bApi"
                ],
                "summary": "\u6211\u7684\u5ba1\u6838",
                "description": "\u6211\u7684\u5ba1\u6838",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content_type",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            }
        },
        "/api/shop/certify/{id}": {
            "get": {
                "tags": [
                    "Shop \u5546\u5e97\u7c7bApi"
                ],
                "summary": "\u7533\u8bf7\u8425\u4e1a\u5ba1\u6279",
                "description": "\u7533\u8bf7\u8425\u4e1a\u5ba1\u6279",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content_type",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Shop \u5546\u5e97\u7c7bApi"
                ],
                "summary": "\u5220\u9664\u8425\u4e1a\u5ba1\u6279\u7533\u8bf7",
                "description": "\u5220\u9664\u8425\u4e1a\u5ba1\u6279\u7533\u8bf7",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content_type",
                        "required": true,
                        "type": "string",
                        "default": "application/json"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f"
                    }
                }
            }
        },
...
    },
...
}